+++ GEI-DS-614G010152122-OP2/src/tpv/productos/Ingrediente.java
@@ -3,7 +3,7 @@ package tpv.productos;
 import tpv.Producto;
 
 public class Ingrediente extends Producto {
-    public Ingrediente(String nombre, String descripcion, double precio, int cantidad) {
+    public Ingrediente(String nombre, String descripcion, double precio, double cantidad) {
         super(nombre, descripcion, precio, cantidad);
     }
 }
